+++ app/Http/Controllers/EntityController.php
@@ -1137,6 +1137,66 @@ class EntityController extends Controller {
         return response()->json(null, 204);
     }
 
+    public function moveEntities(Request $request) {
+        $user = auth()->user();
+        if(!$user->can('entity_write')) {
+            return response()->json([
+                'error' => __('You do not have the permission to modify an entity'),
+            ], 403);
+        }
+        $this->validate($request, [
+            'entity_ids' => 'required|array',
+            'parent_id' => 'nullable|integer|exists:entities,id',
+        ]);
+
+        $entityIds = $request->get('entity_ids');
+        $parentId = $request->get('parent_id');
+
+        foreach($entityIds as $id){
+            if($id == $parentId){
+                return response()->json([
+                    'error' => __('An entity cannot be its own parent'),
+                ], 400);
+            }
+        }
+
+        $parent = null;
+        if(isset($parentId)) {
+            $parent = Entity::find($parentId);
+            if(!isset($parent)) {
+                return response()->json([
+                    'error' => __('The parent entity does not exist'),
+                ], 400);
+            }
+        }
+
+        $startRank = Entity::getNextRank($parentId);
+        $modified = [];
+        DB::beginTransaction();
+        try {
+            for($i = 0; $i < count($entityIds); $i++) {
+                $entity = Entity::findOrFail($entityIds[$i]);
+                $entity->move($parentId, $startRank, $user);
+                $startRank++;
+
+                $modified[] = array(
+                    "entity_id" => $entity->id,
+                    "parent_id" => $entity->root_entity_id,
+                    "rank" => $entity->rank,
+                );
+            }
+        } catch(Exception $e) {
+            DB::rollBack();
+            info("Error while moving Entities:\n$e");
+            return response()->json([
+                'error' => __('An error occurred while moving the entities'),
+            ], 500);
+        }
+
+        DB::commit();
+        return response()->json($modified, 200);
+    }
+
     // DELETE
 
     public function deleteEntity($id) {