+++ resources/js/components/modals/entity/Move.vue
@@ -27,6 +27,20 @@
                     role="form"
                     @submit.prevent="move()"
                 >
+
+                    <div
+                        v-if="isMoveModeMulti"
+                        class="alert alert-warning"
+                    >
+                        {{ t('main.entity.tree.multiedit.warning') }}
+                    </div>
+
+                    <Alert
+                        v-if="invalidSelection"
+                        type="error"
+                        :message="t('main.entity.tree.move.invalid_selection')"
+                    />
+
                     <div
                         v-if="state.hasParent"
                         class="form-check form-switch"
@@ -101,12 +115,19 @@
     import {
         isAllowedSubEntityType,
     } from '@/helpers/helpers.js';
-    
+
     import {
         isPaginated,
     } from '@/helpers/pagination.js';
 
+    import useEntityStore from '../../../bootstrap/stores/entity';
+
+    import Alert from '@/components/Alert.vue';
+
     export default {
+        components: {
+            Alert,
+        },
         props: {
             entity: {
                 required: true,
@@ -120,6 +141,8 @@
                 entity,
             } = toRefs(props);
 
+            const entityStore = useEntityStore();
+
             // FUNCTIONS
             const entitySelected = entity => {
                 state.parent = entity;
@@ -128,22 +151,70 @@
                 if(isPaginated(results)) {
                     results = results.data;
                 }
-                return results.filter(r => {
-                    const onSame = r.id == entity.value.root_entity_id ||
-                        r.id == entity.value.id;
-                    if(onSame) return false;
-                    return isAllowedSubEntityType(r.entity_type_id, entity.value.entity_type_id);
-                });
+
+                if(isMoveModeMulti.value) {
+                    return filterForMultiMove(results);
+                } else {
+                    return results.filter(r => {
+                        const onSame = r.id == entity.value.root_entity_id ||
+                            r.id == entity.value.id;
+                        if(onSame) return false;
+                        return isAllowedSubEntityType(r.entity_type_id, entity.value.entity_type_id);
+                    });
+                }
             };
+
+            const filterForMultiMove = (results) => {
+                if(invalidSelection.value) {
+                    return [];
+                } else {
+                    // As we reqire the same entity type, we can just take the first one
+                    const entity = Object.values(entityStore.treeSelection)[0];
+                    const entities = Object.keys(entityStore.treeSelection).map(id => parseInt(id));
+                    const entityTypeId = entity.entity_type_id;
+
+                    return results.filter(result => {
+                        const onSameEntity = entities.includes(parseInt(result.id));
+                        if(onSameEntity) return false;
+                        return isAllowedSubEntityType(result.entity_type_id, entityTypeId);
+                    });
+                }
+            }
+
             const move = _ => {
                 if(state.dataMissing) {
                     return;
                 }
-                context.emit('confirm', state.moveToRoot ? null : state.parent.id);
+
+                let targets = isMoveModeMulti.value ? Object.keys(entityStore.treeSelection).map(id => parseInt(id)) : [];
+                context.emit('confirm', state.moveToRoot ? null : state.parent.id, targets);
             };
             const closeModal = _ => {
                 context.emit('cancel', false);
             };
+
+            const isMoveModeMulti = computed(() => {
+                return Object.keys(entityStore.treeSelection).length > 1;
+            });
+
+            // We only allow moving entities of the same type
+            const invalidSelection = computed(() => {
+                if(isMoveModeMulti.value) {
+                    let type = null;
+                    const selection = entityStore.treeSelection;
+                    for(let index in selection) {
+                        const item = selection[index];
+                        const entityTypeId = item?.entity_type_id;
+                        if(type == null) {
+                            type = entityTypeId;
+                        } else if(type !== entityTypeId) {
+                            return true;
+                        }
+                    }
+
+                    return false;
+                } else return false;
+            });
 
             // DATA
             const state = reactive({
@@ -161,6 +232,8 @@
                 // HELPERS
                 searchEntity,
                 // LOCAL
+                invalidSelection,
+                isMoveModeMulti,
                 entitySelected,
                 filterEntityResults,
                 move,