+++ resources/js/helpers/modal.js
@@ -571,13 +571,25 @@ export function ShowMoveEntity(entity, onMoved) {
             onCancel(e) {
                 modal.destroy();
             },
-            onConfirm(parentId) {
-                useEntityStore().move(entity.id, parentId).then(data => {
-                    if(!!onMoved) {
-                        onMoved(entity.id, parentId, data);
-                    }
-                    modal.destroy();
-                });
+            onConfirm(parentId, entity_ids) {
+                if(entity_ids && entity_ids.length > 0) {
+                    useEntityStore().moveMultiple(entity_ids, parentId).then(data => {
+                        if(!!onMoved) {
+                            entity_ids.forEach(entity_id => {
+                                onMoved(entity_id, parentId, data);
+                            });
+                        }
+                        useEntityStore().setTreeSelectionMode(false);
+                        modal.destroy();
+                    });
+                } else {
+                    useEntityStore().move(entity.id, parentId).then(data => {
+                        if(!!onMoved) {
+                            onMoved(entity.id, parentId, data);
+                        }
+                        modal.destroy();
+                    });
+                }
             },
         },
     });